status bar integrated with redux
