+++ app/Components/Wizard.js
@@ -1,26 +1,28 @@
 import React, { PropTypes } from 'react';
-import StatusBar from './StatusBar';
+import StatusBarContainer  from './StatusBar';
 import Steps from './steps'
+import { connect } from 'react-redux'
+import { nextStep, prevStep } from '../actions'
 
-const Wizard = ( { currentStatus, onNextStep, onPrevStep } ) => {
+const Wizard = ( { currentStep, onNextStep, onPrevStep, statusBar} ) => {
 
-    const stepCount =  Steps.length,
+    const maxSteps =  Steps.length,
           hidden = {
                 display: 'none'
           };
 
     const isFinalStep = () => {
-        return stepCount === currentStatus;
+        return maxSteps === currentStep;
     }
 
     const isFirstStep = () => {
-        return currentStatus === 1
+        return currentStep === 1
     }
 
     return (
         <div>
-            { Steps[currentStatus - 1].component }
-            <StatusBar currentStatus={ currentStatus } maxStatus={ stepCount } />
+            { Steps[currentStep - 1].component }
+            { statusBar ?  <StatusBarContainer maxSteps={ maxSteps } /> : '' }
             <button onClick={ () => { onPrevStep() } } style={ isFirstStep() ? hidden : {} }>Back</button>
             <button onClick={ () => { onNextStep() } } style={ isFinalStep() ? hidden : {} }>Next</button>
         </div>
@@ -31,8 +33,30 @@ Wizard.PropTypes = {
     wizardStep: PropTypes.number.isRequired,
     steps : PropTypes.array.isRequired,
     onNextStep : PropTypes.func.isRequired,
-    onPrevStep : PropTypes.func.isRequired
+    onPrevStep : PropTypes.func.isRequired,
 }
 
-export default Wizard;
+const mapStateToProps = (state, ownProps) => {
+    return {
+        currentStep: state.wizard,
+        statusBar: ownProps.statusBar || true
+    }
+}
+
+const mapDispatchToProps = (dispatch) => {
+    return {
+        onNextStep: () => {
+            dispatch(nextStep())
+        },
+        onPrevStep: () => {
+            dispatch(prevStep())
+        }
+    }
+}
+
+const WizardContainer = connect(mapStateToProps, mapDispatchToProps)(Wizard);
+
+export default WizardContainer;
+
+
